requirements.txt
gevent==0.13.7
mechanize==0.2.5
pyyaml==3.10
requests==1.2.3
